fixing memory and speeding up parallel join

+++ include/kProcessor/kDataFrame.hpp
@@ -752,7 +752,7 @@ private:
     kmerDecoder * KD;
 public:
     typedef typename phmap::parallel_flat_hash_map<std::uint64_t,
-    std::uint64_t,
+    std::uint32_t,
     std::hash<uint64_t>,
     std::equal_to<uint64_t>,
     std::allocator<std::pair<const uint64_t, uint64_t>>,
@@ -790,7 +790,7 @@ public:
 class kDataFramePHMAP : public kDataFrame {
 public:
     typedef  phmap::parallel_flat_hash_map<std::uint64_t,
-    std::uint64_t,
+    std::uint32_t,
     std::hash<uint64_t>,
     std::equal_to<uint64_t>,
     std::allocator<std::pair<const uint64_t, uint64_t>>,
